requirements.txt
requests
mysql-connector-python
fake-useragent